project: move ItemProvider into SharedPocketKit
