+++ PocketKit/Sources/SaveToPocketKit/ExtensionContext.swift
@@ -3,12 +3,14 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import Foundation
+import SharedPocketKit
 
 protocol ExtensionContext {
     var extensionItems: [ExtensionItem] { get }
 
     func completeRequest(returningItems items: [Any]?, completionHandler: ((Bool) -> Void)?)
 }
+
 extension NSExtensionContext: ExtensionContext {
     var extensionItems: [ExtensionItem] {
         inputItems.compactMap { $0 as? ExtensionItem }
@@ -23,13 +25,3 @@ extension NSExtensionItem: ExtensionItem {
         attachments?.compactMap { $0 as ItemProvider }
     }
 }
-
-protocol ItemProvider {
-    func hasItemConformingToTypeIdentifier(_ typeIdentifier: String) -> Bool
-
-    func loadItem(
-        forTypeIdentifier typeIdentifier: String,
-        options: [AnyHashable: Any]?
-    ) async throws -> NSSecureCoding
-}
-extension NSItemProvider: ItemProvider { }
