+++ PocketKit/Sources/SharedPocketKit/ItemProvider.swift
@@ -1,0 +1,16 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import Foundation
+
+public protocol ItemProvider {
+    func hasItemConformingToTypeIdentifier(_ typeIdentifier: String) -> Bool
+
+    func loadItem(
+        forTypeIdentifier typeIdentifier: String,
+        options: [AnyHashable: Any]?
+    ) async throws -> NSSecureCoding
+}
+
+extension NSItemProvider: ItemProvider { }
